+++ packages/liveblocks-react-ui/src/components/Thread.tsx
@@ -1,8 +1,14 @@
 "use client";
 
-import type { BaseMetadata, CommentData, ThreadData } from "@liveblocks/core";
+import type {
+  BaseMetadata,
+  CommentData,
+  DM,
+  ThreadData,
+} from "@liveblocks/core";
 import {
-  useEditThreadMetadata,
+  useMarkThreadAsResolved,
+  useMarkThreadAsUnresolved,
   useThreadSubscription,
 } from "@liveblocks/react";
 import * as TogglePrimitive from "@radix-ui/react-toggle";
@@ -31,7 +37,6 @@ import type {
   ThreadOverrides,
 } from "../overrides";
 import { useOverrides } from "../overrides";
-import type { ThreadMetadata } from "../types";
 import { classNames } from "../utils/class-names";
 import { findLastIndex } from "../utils/find-last-index";
 import type { CommentProps } from "./Comment";
@@ -40,7 +45,7 @@ import { Composer } from "./Composer";
 import { Button } from "./internal/Button";
 import { Tooltip, TooltipProvider } from "./internal/Tooltip";
 
-export interface ThreadProps<M extends BaseMetadata = ThreadMetadata>
+export interface ThreadProps<M extends BaseMetadata = DM>
   extends ComponentPropsWithoutRef<"div"> {
   /**
    * The thread to display.
@@ -128,7 +133,7 @@ export interface ThreadProps<M extends BaseMetadata = ThreadMetadata>
  * </>
  */
 export const Thread = forwardRef(
-  <M extends BaseMetadata = ThreadMetadata>(
+  <M extends BaseMetadata = DM>(
     {
       thread,
       indentCommentContent = true,
@@ -149,7 +154,8 @@ export const Thread = forwardRef(
     }: ThreadProps<M>,
     forwardedRef: ForwardedRef<HTMLDivElement>
   ) => {
-    const editThreadMetadata = useEditThreadMetadata();
+    const markThreadAsResolved = useMarkThreadAsResolved();
+    const markThreadAsUnresolved = useMarkThreadAsUnresolved();
     const $ = useOverrides(overrides);
     const firstCommentIndex = useMemo(() => {
       return showDeletedComments
@@ -212,9 +218,18 @@ export const Thread = forwardRef(
       (resolved: boolean) => {
         onResolvedChange?.(resolved);
 
-        editThreadMetadata({ threadId: thread.id, metadata: { resolved } });
+        if (resolved) {
+          markThreadAsResolved(thread.id);
+        } else {
+          markThreadAsUnresolved(thread.id);
+        }
       },
-      [editThreadMetadata, onResolvedChange, thread.id]
+      [
+        markThreadAsResolved,
+        markThreadAsUnresolved,
+        onResolvedChange,
+        thread.id,
+      ]
     );
 
     const handleCommentDelete = useCallback(
@@ -240,9 +255,7 @@ export const Thread = forwardRef(
             showActions === "hover" && "lb-thread:show-actions-hover",
             className
           )}
-          data-resolved={
-            (thread.metadata as ThreadMetadata).resolved ? "" : undefined
-          }
+          data-resolved={thread.resolved ? "" : undefined}
           data-unread={unreadIndex !== undefined ? "" : undefined}
           dir={$.dir}
           {...props}
@@ -280,13 +293,13 @@ export const Thread = forwardRef(
                     isFirstComment && showResolveAction ? (
                       <Tooltip
                         content={
-                          (thread.metadata as ThreadMetadata).resolved
+                          thread.resolved
                             ? $.THREAD_UNRESOLVE
                             : $.THREAD_RESOLVE
                         }
                       >
                         <TogglePrimitive.Root
-                          pressed={(thread.metadata as ThreadMetadata).resolved}
+                          pressed={thread.resolved}
                           onPressedChange={handleResolvedChange}
                           asChild
                         >
@@ -294,12 +307,12 @@ export const Thread = forwardRef(
                             className="lb-comment-action"
                             onClick={stopPropagation}
                             aria-label={
-                              (thread.metadata as ThreadMetadata).resolved
+                              thread.resolved
                                 ? $.THREAD_UNRESOLVE
                                 : $.THREAD_RESOLVE
                             }
                           >
-                            {(thread.metadata as ThreadMetadata).resolved ? (
+                            {thread.resolved ? (
                               <ResolvedIcon className="lb-button-icon" />
                             ) : (
                               <ResolveIcon className="lb-button-icon" />
@@ -347,6 +360,6 @@ export const Thread = forwardRef(
       </TooltipProvider>
     );
   }
-) as <M extends BaseMetadata = ThreadMetadata>(
+) as <M extends BaseMetadata = DM>(
   props: ThreadProps<M> & RefAttributes<HTMLDivElement>
 ) => JSX.Element;