+++ packages/liveblocks-react/src/index.ts
@@ -40,6 +40,8 @@ export {
   useDeleteThread,
   useEditComment,
   useEditThreadMetadata,
+  useMarkThreadAsResolved,
+  useMarkThreadAsUnresolved,
   useErrorListener,
   useEventListener,
   useHistory,