+++ packages/liveblocks-react/src/types.ts
@@ -53,6 +53,12 @@ export type UseThreadsOptions<M extends BaseMetadata> = {
    */
   query?: {
     /**
+     * Whether to only return threads marked as resolved or unresolved. If not provided,
+     * all threads will be returned.
+     */
+    resolved?: boolean;
+
+    /**
      * The metadata to filter the threads by. If provided, only threads with metadata that matches
      * the provided metadata will be returned. If not provided, all threads will be returned.
      */
@@ -779,6 +785,24 @@ type RoomContextBundleCommon<
    * editThreadMetadata({ threadId: "th_xxx", metadata: {} })
    */
   useEditThreadMetadata(): (options: EditThreadMetadataOptions<M>) => void;
+
+  /**
+   * Returns a function that marks a thread as resolved.
+   *
+   * @example
+   * const markThreadAsResolved = useMarkThreadAsResolved();
+   * markThreadAsResolved("th_xxx");
+   */
+  useMarkThreadAsResolved(): (threadId: string) => void;
+
+  /**
+   * Returns a function that marks a thread as unresolved.
+   *
+   * @example
+   * const markThreadAsUnresolved = useMarkThreadAsUnresolved();
+   * markThreadAsUnresolved("th_xxx");
+   */
+  useMarkThreadAsUnresolved(): (threadId: string) => void;
 
   /**
    * Returns a function that adds a comment to a thread.