+++ CHANGELOG.md
@@ -1,17 +1,25 @@
-## v2.2.0 (not published yet)
+## v2.2.0
+
+We are making `resolved` a first-class citizen property on
+[threads](https://liveblocks.io/docs/products/comments/concepts#Threads), for
+more information about this change please read our
+[Upgrade Guide for 2.2](https://liveblocks.io/docs/guides/upgrading/2.2).
 
 ### `@liveblocks/react`
 
+- Add `useMarkThreadAsResolved` and `useMarkThreadAsUnresolved` hooks.
+- Support `query.resolved` when filtering threads.
 - The
   [`useStorageStatus`](https://liveblocks.io/docs/api-reference/liveblocks-react#useStorageStatus)
   hook now also has a `{ smooth: true }` setting to make building calm UIs with
-  it a bit easier
+  it a bit easier.
 - The `useClient()` hook is now also available for users of
   `createRoomContext()` and/or `createLiveblocksContext()`
 - Fix: avoid unnecessary re-renders if inbox notifications haven't changed
 
 ### `@liveblocks/react-ui`
 
+- Use first-class citizen `resolved` property in `Thread` component.
 - Preserve rich text when pasting into the composer.
 - Add support for custom links to the composer. (either by pasting URLs with
   plain text selected or by pasting existing links)
@@ -20,6 +28,13 @@
   focused.
 - Fix improper `useTransition` fallback which would break on React versions
   lower than 18.
+
+### `@liveblocks/node`
+
+- Add `markThreadAsResolved` and `markThreadAsUnresolved` methods.
+- Add `ThreadMarkedAsResolvedEvent` and `ThreadMarkedAsUnresolvedEvent` webhook
+  events.
+- Support `query.resolved` when querying threads.
 
 ### @liveblocks/react-lexical